+++ examples/demo_canonical_fastq.rs
@@ -10,9 +10,7 @@ fn main() {
         "examples/example.fastq".to_owned()
     };
 
-    let fastq_reader = FastQReader::from_filename(filename)
-        .unwrap_or_die("FastQ file error!")
-        .flatten();
+    let fastq_reader = FastQReader::from_filename(filename).unwrap_or_fail().flatten();
 
     for mut record in fastq_reader {
         record.recode_dna_reads();